pyproject.toml
[project]
name = "rewrite-mate"
version = "0.1.0"
description = ""
authors = [
    {name = "Cha Hwa Young",email = "chahwayoung214@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "langchain (>=0.3.27,<0.4.0)",
    "pytest (>=8.4.1,<9.0.0)",
    "langchain-anthropic (>=0.3.18,<0.4.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "streamlit (>=1.47.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
